+++ lib/coloraide/spaces/a98_rgb.py
@@ -5,7 +5,6 @@ from .srgb import SRGB
 from .xyz import XYZ
 from .. import util
 import re
-import math
 
 
 def lin_a98rgb_to_xyz(rgb):
@@ -42,13 +41,13 @@ def xyz_to_lin_a98rgb(xyz):
 def lin_a98rgb(rgb):
     """Convert an array of a98-rgb values in the range 0.0 - 1.0 to linear light (un-corrected) form."""
 
-    return [math.copysign(abs(val) ** (563 / 256), val) for val in rgb]
+    return [util.npow(val, 563 / 256) for val in rgb]
 
 
 def gam_a98rgb(rgb):
     """Convert an array of linear-light a98-rgb  in the range 0.0-1.0 to gamma corrected form."""
 
-    return [math.copysign(abs(val) ** (256 / 563), val) for val in rgb]
+    return [util.npow(val, 256 / 563) for val in rgb]
 
 
 class A98RGB(SRGB):