+++ lib/coloraide/spaces/hwb.py
@@ -46,7 +46,7 @@ class HWB(Cylindrical, Space):
     GAMUT_CHECK = "hsl"
     WHITE = _cat.WHITES["D65"]
 
-    _range = (
+    RANGE = (
         GamutBound([Angle(0.0), Angle(360.0)]),
         GamutBound([Percent(0.0), Percent(100.0)]),
         GamutBound([Percent(0.0), Percent(100.0)])