+++ lib/coloraide/spaces/jzczhz.py
@@ -59,7 +59,7 @@ class JzCzhz(Cylindrical, Space):
     DEFAULT_MATCH = re.compile(RE_DEFAULT_MATCH.format(color_space=SPACE))
     WHITE = _cat.WHITES["D65"]
 
-    _range = (
+    RANGE = (
         GamutUnbound([0.0, 1.0]),
         GamutUnbound([0.0, 1.0]),
         GamutUnbound([Angle(0.0), Angle(360.0)]),